python-fastapi/requirements.txt
anthropic>=0.25.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
python-docx>=1.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
